make Helper.py mostly circular import safe
