+++ src/hooks/Helpers.py
@@ -1,8 +1,5 @@
 from typing import Optional
-from BaseClasses import MultiWorld
-from ..Locations import ManualLocation
-from ..Items import ManualItem
-
+from BaseClasses import MultiWorld, Item, Location
 
 # Use this if you want to override the default behavior of is_option_enabled
 # Return True to enable the category, False to disable it, or None to use the default behavior
@@ -11,10 +8,10 @@ def before_is_category_enabled(multiworld: MultiWorld, player: int, category_nam
 
 # Use this if you want to override the default behavior of is_option_enabled
 # Return True to enable the item, False to disable it, or None to use the default behavior
-def before_is_item_enabled(multiworld: MultiWorld, player: int, item: ManualItem) -> Optional[bool]:
+def before_is_item_enabled(multiworld: MultiWorld, player: int, item: Item) -> Optional[bool]:
     return None
 
 # Use this if you want to override the default behavior of is_option_enabled
 # Return True to enable the location, False to disable it, or None to use the default behavior
-def before_is_location_enabled(multiworld: MultiWorld, player: int, location: ManualLocation) -> Optional[bool]:
+def before_is_location_enabled(multiworld: MultiWorld, player: int, location: Location) -> Optional[bool]:
     return None
